Move backoff functions to retry file and use generic getCode function
